arrow roughly works - points towards attention point, and sits on FAKE eye point; adding data recordings to the gitignore file

+++ Assets/AttentionEventsController.cs
@@ -13,7 +13,7 @@ public class AttentionEventsController : MonoBehaviour {
 	[System.Serializable]
 	public class AttentionEvent
 	{
-		public int startTime;
+		public float startTime;
 		public float hAngle;
 		public float vAngle;
 		public string type;
